+++ frontend/src/views/node/Overview.vue
@@ -5,315 +5,291 @@ file, You can obtain one at http://mozilla.org/MPL/2.0/.
 -->
 <template>
   <div class="flex flex-col gap-2">
-    <div class="px-3 py-2 mb-2">
-      <t-breadcrumbs>{{ $route.params.node }} Overview</t-breadcrumbs>
+    <div class="px-3 py-2">
+      <t-breadcrumbs>{{ $route.params.node }}</t-breadcrumbs>
     </div>
-    <div class="p-4 border rounded-md border-talos-gray-300 dark:border-talos-gray-600 flex-1 flex flex-col overflow-hidden">
-      <div class="flex flex-1 gap-2">
-        <div class="flex-1">
-          <t-chart
-             class="h-full"
-             name="cpu"
-             title="CPU Utilization"
-             type="area"
-             talos
-             :resource="{type: 'CPUStat', namespace: 'perf', tail_events: 25}"
-             :context="context"
-             :point-fn="handleCPU"
-          />
-        </div>
-        <div class="flex-1">
-          <t-chart
-             class="h-full"
-             name="mem"
-             title="Memory Usage"
-             type="area"
-             talos
-             :resource="{type: 'MemoryStat', namespace: 'perf', tail_events: 25}"
-             :context="context"
-             :point-fn="handleMem"
-          />
-        </div>
-      </div>
-      <div class="flex flex-1 gap-2">
-        <div class="flex-1">
-          <t-chart
-             class="h-full"
-             name="procs"
-             title="Processes"
-             type="area"
-             talos
-             :resource="{type: 'CPUStat', namespace: 'perf', tail_events: 25}"
-             :context="context"
-             :point-fn="handleProcs"
-          />
-        </div>
-      </div>
+    <div class="px-3">
+      <h1 class="text-lg tracking-tight text-talos-gray-900 dark:text-white font-bold">Node Overview</h1>
     </div>
-    <h4>Process List</h4>
-    <div class="p-2 text-xs border rounded-md border-talos-gray-300 dark:border-talos-gray-600 flex-1 flex flex-col overflow-hidden">
-      <div class="grid grid-cols-12 uppercase font-bold select-none">
-        <div v-for="h in headers" @click="() => sortBy(h.id)" :key="h.id" class="flex flex-row items-center cursor-pointer gap-1">
-          <span>{{ h.header || h.id }}</span>
-          <arrow-down-icon class="w-3 h-3" :class="{ transform: sortReverse, 'rotate-180': sortReverse }" v-if="sort === h.id"/>
-        </div>
+    <div>
+      <t-alert v-if="err" title="Failed to Fetch Node Summary" type="error">
+        {{ err }}.
+      </t-alert>
+      <div v-else-if="!item" class="flex flex-row justify-center items-center w-full h-full">
+        <t-spinner/>
       </div>
-      <div class="flex-1 overflow-x-auto">
-        <div class="grid grid-cols-12" v-for="process in processes" :key="process.pid">
-          <div>
-            {{ process.pid }}
-          </div>
-          <div>
-            {{ process.state }}
-          </div>
-          <div>
-            {{ process.threads }}
-          </div>
-          <div>
-            {{ process.cpu.toFixed(1) }}
-          </div>
-          <div>
-            {{ process.mem.toFixed(1) }}
-          </div>
-          <div>
-            {{ humanizeBytes(process.virtualMemory) }}
-          </div>
-          <div>
-            {{ humanizeBytes(process.residentMemory) }}
-          </div>
-          <div>
-            {{ process.cpuTime }}
-          </div>
-          <div class="col-span-4 truncate">
-            {{ process.command }} {{ process.args }}
-          </div>
-        </div>
+      <div v-else class="flex flex-row gap-3">
+        <stacked-list :items="summary" class="flex-1">
+          <template v-slot:header>
+            <div class="font-medium px-3 py-2">
+             Kubernetes
+            </div>
+          </template>
+          <template v-slot:default="slot">
+            <div class="items-center grid grid-cols-3 p-3 h-9">
+              <div class="truncate">
+                {{ slot.item.name }}
+              </div>
+              <div class="truncate col-span-2">
+                <div class="flex gap-1" v-if="slot.item.name === 'Roles'">
+                  <t-label v-for="role in slot.item.value" v-bind:key="role" class="inline text-sm">{{ role }}</t-label>
+                </div>
+                <div class="flex gap-1" v-else-if="slot.item.name === 'Status'">
+                  <t-label :col="slot.item.value === 'ready' ? 'green' : 'red'" class="uppercase">
+                    <template v-slot:icon>
+                      <check-circle-icon class="w-5 h-5" v-if="slot.item.value === 'ready'"/>
+                      <exclamation-circle-icon class="w-5 h-5" v-else/>
+                    </template>
+                    {{ slot.item.value }}
+                  </t-label>
+                </div>
+                <template v-else>
+                  {{ slot.item.value }}
+                </template>
+              </div>
+            </div>
+          </template>
+        </stacked-list>
+        <stacked-list :items="nodeInfo" class="flex-1">
+          <template v-slot:header>
+            <div class="font-medium px-3 py-2">
+              Node
+            </div>
+          </template>
+          <template v-slot:default="slot">
+            <div class="items-center grid grid-cols-3 p-3 h-9">
+              <div class="truncate">
+                {{ slot.item.name }}
+              </div>
+              <div class="truncate col-span-2">
+                {{ slot.item.value }}
+              </div>
+            </div>
+          </template>
+        </stacked-list>
       </div>
     </div>
+    <watch
+        class="flex-1"
+        :resource="{type: 'services'}"
+        showCount
+        itemName="Service"
+        talos
+        search="Search service by name"
+        :context="getContext()">
+      <template v-slot:header>
+        <div class="flex items-center grid grid-cols-8">
+          <div class="col-span-6 block">
+            ID
+          </div>
+          <div class="block">
+            Running
+          </div>
+          <div class="block">
+            Health
+          </div>
+        </div>
+      </template>
+      <template v-slot:default="slot">
+        <a
+          class="block hover:bg-talos-gray-50 dark:hover:bg-talos-gray-800"
+          >
+          <div class="flex items-center px-3 py-2 grid grid-cols-8 gap-4">
+            <div class="col-span-6 block">
+              <p
+                class="text-sm font-medium truncate text-talos-gray-900 dark:text-talos-gray-100"
+                >
+                {{ slot.item.metadata.id }}
+              </p>
+            </div>
+            <div class="block justify-center">
+              <p
+                class="text-sm font-medium truncate text-talos-gray-900 dark:text-talos-gray-100"
+                >
+                <check-circle-icon class="ok" v-if="slot.item.spec.running"/>
+                <x-circle-icon class="error" v-else/>
+              </p>
+            </div>
+            <div class="block justify-center">
+              <p
+                class="text-sm font-medium truncate text-talos-gray-900 dark:text-talos-gray-100"
+                >
+                <check-circle-icon class="ok" v-if="slot.item.spec.healthy"/>
+                <question-mark-circle-icon class="info" v-else-if="slot.item.spec.unknown"/>
+                <exclamation-circle-icon class="error" v-else/>
+              </p>
+            </div>
+          </div>
+        </a>
+      </template>
+    </watch>
   </div>
 </template>
 
 <script type="ts">
-import { getContext } from '../../context';
 import { Runtime } from '../../api/common/theila.pb';
-import { ref, onMounted, onUnmounted, computed } from 'vue';
-import { MachineService } from '../../api/grpc';
+import { getContext } from '../../context';
+import { ref, onMounted, computed } from 'vue';
+import { ResourceService } from '../../api/grpc';
 import TBreadcrumbs from '../../components/TBreadcrumbs.vue';
-import TChart from '../../components/TChart.vue';
+import TSpinner from '../../components/TSpinner.vue';
+import TLabel from '../../components/TLabel.vue';
+import TAlert from '../../components/TAlert.vue';
+import StackedList from '../../components/StackedList.vue';
 import {
-  ArrowDownIcon,
+  CheckCircleIcon,
+  ExclamationCircleIcon,
+  QuestionMarkCircleIcon,
+  XCircleIcon,
 } from '@heroicons/vue/solid';
-
-function humanizeBytes(size) {
-  var gb = Math.pow(1024, 3);
-  var mb = Math.pow(1024, 2);
-  var kb = 1024;
-
-  if (size >= gb)
-    return Math.floor(size / gb) + ' GB';
-  else if (size >= mb)
-    return Math.floor(size / mb) + ' MB';
-  else if (size >= kb)
-    return Math.floor(size / kb) + ' KB';
-  else
-    return size + ' B';
-}
-
-function diff(a, b) {
-  const result = {};
-
-  for(const key in a) {
-    const left = a[key];
-    const right = b[key];
-
-    const tleft = typeof left;
-    const tright = typeof right;
-
-    if(tleft !== tright) {
-      continue;
-    }
-
-    if(tleft === "object")
-      result[key] = diff(left, right);
-    else
-      result[key] = left - right;
-  }
-
-  return result;
-}
+import Watch from '../../components/Watch.vue';
 
 export default {
   components: {
+    CheckCircleIcon,
+    ExclamationCircleIcon,
     TBreadcrumbs,
-    TChart,
-    ArrowDownIcon,
+    TSpinner,
+    TLabel,
+    TAlert,
+    StackedList,
+    QuestionMarkCircleIcon,
+    XCircleIcon,
+    Watch,
   },
 
   setup() {
-    const processes = ref([]);
+    const item = ref(null);
     const context = getContext();
-    const headers = [
-      { id: "pid" },
-      { id: "state" },
-      { id: "threads" },
-      { id: "cpu", header: "CPU %" },
-      { id: "mem", header: "Memory %" },
-      { id: "virtualMemory", header: "Virt Memory" },
-      { id: "residentMemory", header: "Res Memory" },
-      { id: "cpuTime", header: "Time+" },
-      { id: "command" },
-    ];
-    const sort = ref("cpu");
-    const sortReverse = ref(true);
+    const err = ref(null);
 
-    let memTotal = 0;
-    let interval;
-    
-    const sum = (obj, ...args) => {
-      let res = 0;
+    onMounted(async () => {
+      try {
+        const hostname = await ResourceService.Get({
+          type: "hostname",
+          id: "hostname",
+        }, {
+          runtime: Runtime.Talos,
+          context: context,
+        });
+
+        const response = await ResourceService.Get({
+          namespace: "default",
+          type: "nodes.v1",
+          id: hostname.spec.hostname,
+        }, {
+          runtime: Runtime.Kubernetes,
+          context: context,
+        });
+
+        item.value = response;
+      } catch(e) {
+        err.value = e.toString();
+      }
+    });
+
+    const getField = (...args) => {
+      if(item.value == null)
+        return null;
+
+      let res = item.value;
       for(const k of args) {
-        res += obj[k] || 0;
+        if(!res[k])
+          return null;
+
+        res = res[k];
       }
 
       return res;
     };
 
-    const getCPUTotal = (stat) => {
-      const idle = sum(stat, "idle", "iowait");
-      const nonIdle = sum(stat, "user", "nice", "system", "irq", "steal", "softIrq");
+    const os = computed(() => {
+      return getField("status", "nodeInfo", "osImage") || "";
+    });
 
-      return idle + nonIdle;
-    };
+    const ip = computed(() => {
+      const addresses = getField("status", "addresses");
+      if(!addresses)
+        return "unknown";
 
-    onMounted(() => {
-      const prevProcs = {};
-      let prevCPU = 0;
-
-      interval = setInterval(async () => {
-        if(memTotal === 0)
-          return;
-
-        const resp = await MachineService.Processes({}, {
-          runtime: Runtime.Talos,
-          context: context,
-        });
-
-        let procs = [];
-
-        const r = await MachineService.SystemStat({}, {
-          runtime: Runtime.Talos,
-          context: context,
-        });
-
-        const systemStat = r.messages[0];
-        const cpuTotal = getCPUTotal(systemStat.cpuTotal) / systemStat.cpu.length;
-
-        const total = memTotal * 1024;
-
-        for(const message of resp.messages) {
-          for(const proc of message.processes) {
-            let cpuDiff = 0;
-
-            if(prevProcs[proc.pid] && proc.cpuTime) {
-              cpuDiff = proc.cpuTime - prevProcs[proc.pid].cpuTime;
-            }
-
-            procs.push({
-              mem: (proc.residentMemory || 0) / total * 100,
-              cpu: cpuDiff / (cpuTotal - prevCPU) * 100,
-              threads: proc.threads,
-              pid: proc.pid,
-              state: proc.state,
-              virtualMemory: parseInt(proc.virtualMemory || 0),
-              residentMemory: parseInt(proc.residentMemory || 0),
-              command: proc.command,
-              cpuTime: proc.cpuTime || 0,
-              args: proc.args,
-            });
-
-            prevProcs[proc.pid] = proc;
-          }
+      let addr;
+      for(const a of addresses) {
+        if (a.type == "InternalIP") {
+          addr = a.address;
         }
 
-        prevCPU = cpuTotal;
+        if (a.type == "ExternalIP") {
+          addr = a.address;
+          break;
+        }
+      }
 
-        processes.value = procs;
-      }, 2000);
+      return addr;
     });
 
-    onUnmounted(() => {
-      clearInterval(interval);
+    const status = computed(() => {
+      const conditions = getField("status", "conditions");
+      if(!conditions)
+        return "not ready";
+
+      for(const c of conditions) {
+        if(c["type"] === "Ready" && c["status"] === "True")
+          return "ready";
+      }
+
+      return "not ready";
     });
 
-    const handleCPU = (oldObj, newObj) => {
-      const delta = diff(oldObj, newObj);
-      const stat = delta["cpuTotal"];
-      const total = getCPUTotal(stat);
+    const roles = computed(() => {
+      const roles = [];
 
-      return {
-        system: stat["system"]/total * 100,
-        user: stat["user"]/total * 100,
-      };
-    };
-
-    const handleMem = (_, m) => {
-      const used = m["used"] - m["cached"] - m["buffers"]
-
-      memTotal = m["total"];
-
-      return {
-        total: used/m["total"] * 100,
+      for (const label in item.value.metadata.labels) {
+        if (label.indexOf("node-role.kubernetes.io/") != -1) {
+          roles.push(label.split("/")[1]);
+        }
       }
-    };
 
-    const handleProcs = (oldObj, newObj) => {
-      const delta = diff(oldObj, newObj);
+      return roles;
+    });
 
-      return {
-        created: delta["processCreated"],
-        running: newObj["processRunning"],
-        blocked: newObj["processBlocked"],
-      }
-    };
+    const kubeletVersion = computed(() => {
+      return getField("status", "nodeInfo", "kubeletVersion") || "";
+    });
 
     return {
-      handleMem,
-      handleCPU,
-      handleProcs,
-      context,
-      sort,
-      headers,
-      sortReverse,
-      humanizeBytes,
-      processes: computed(() => {
-        return [].concat(processes.value).sort((a, b) => {
-          let res = 0;
-          if(a[sort.value] > b[sort.value]) {
-            res = 1;
-          } else if(a[sort.value] < b[sort.value]) {
-            res = -1;
-          }
-
-          return sortReverse.value ? -1 * res : res;
-        });
+      item,
+      getContext,
+      err,
+      summary: computed(() => {
+        return [
+          {name: "Hostname", value: getField("metadata", "name")},
+          {name: "Kubelet Version", value: kubeletVersion.value},
+          {name: "Kube Proxy Version", value: getField("status", "nodeInfo", "kubeProxyVersion")},
+          {name: "Roles", value: roles.value},
+          {name: "Status", value: status.value},
+        ];
       }),
-      sortBy: (id) => {
-        if(id === sort.value)
-          sortReverse.value = !sortReverse.value
-        else
-          sortReverse.value = true;
-
-        sort.value = id;
-      }
+      nodeInfo: computed(() => {
+        return [
+          {name: "IP", value: ip.value},
+          {name: "OS", value: os.value},
+          {name: "Architecture", value: getField("status", "nodeInfo", "architecture")},
+          {name: "Kernel Version", value: getField("status", "nodeInfo", "kernelVersion")},
+          {name: "Container Runtime Version", value: getField("status", "nodeInfo", "containerRuntimeVersion")},
+        ];
+      }),
     }
   }
 };
 </script>
+
 <style scoped>
 .ok {
   @apply h-5 w-5 text-green-400;
 }
-
+.info {
+  @apply h-5 w-5 text-blue-400;
+}
 .error {
   @apply h-5 w-5 text-red-600;
 }
